ci: unregress tart image gen for macos
